package renamed, groupService created
