+++ pom.xml
@@ -8,16 +8,16 @@
     <version>2.1.6.RELEASE</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
-  <groupId>com.student.enterprise</groupId>
+  <groupId>student.enterprise</groupId>
   <artifactId>project</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>sep</name>
-  <description>Demo project for Spring Boot</description>
+  <version>1.0</version>
+  <name>SEP</name>
 
   <properties>
     <java.version>1.8</java.version>
 		<lombok.version>1.18.10</lombok.version>
     <guava.version>28.1-jre</guava.version>
+    <mockito.version>3.1.0</mockito.version>
   </properties>
 
   <dependencies>
@@ -46,6 +46,12 @@
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
       <version>${guava.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-core</artifactId>
+      <version>${mockito.version}</version>
+      <scope>test</scope>
     </dependency>
