--- src/main/java/com/student/enterprise/project/SEPApplication.java
@@ -1,13 +1,0 @@
-package com.student.enterprise.project;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SEPApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SEPApplication.class, args);
-	}
-
-}
