--- src/main/java/com/student/enterprise/project/configuration/SepConfig.java
@@ -1,8 +1,0 @@
-package com.student.enterprise.project.configuration;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SepConfig {
-
-}
