--- src/main/java/com/student/enterprise/project/converter/jpa/LocalDateTimeAttributeConverter.java
@@ -1,26 +1,0 @@
-package com.student.enterprise.project.converter.jpa;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
-
-@Converter
-public class LocalDateTimeAttributeConverter implements AttributeConverter<LocalDateTime, Timestamp> {
-
-  @Override
-  public Timestamp convertToDatabaseColumn(LocalDateTime localDateTime) {
-    if (localDateTime == null) {
-      return null;
-    }
-    return Timestamp.valueOf(localDateTime);
-  }
-
-  @Override
-  public LocalDateTime convertToEntityAttribute(Timestamp timestamp) {
-    if (timestamp == null) {
-      return null;
-    }
-    return timestamp.toLocalDateTime();
-  }
-}
