--- src/main/java/com/student/enterprise/project/dto/enums/UserRole.java
@@ -1,7 +1,0 @@
-package com.student.enterprise.project.dto.enums;
-
-public enum UserRole {
-  USER,
-  MODERATOR,
-  OWNER
-}
