--- src/main/java/com/student/enterprise/project/entity/UserRoleEntity.java
@@ -1,36 +1,0 @@
-package com.student.enterprise.project.entity;
-
-import com.student.enterprise.project.dto.enums.UserRole;
-import java.io.Serializable;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-/*
-  Enum entity that defines role
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "user_roles")
-public class UserRoleEntity implements Serializable {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Enumerated(EnumType.STRING)
-  @Column(name = "role_name")
-  private UserRole roleName;
-
-}
