--- src/main/java/com/student/enterprise/project/exception/ConverterException.java
@@ -1,13 +1,0 @@
-package com.student.enterprise.project.exception;
-
-public class ConverterException extends RuntimeException {
-
-  public ConverterException(String errorMessage) {
-    super(errorMessage);
-  }
-
-  public ConverterException(String errorMessage, Throwable ex) {
-    super(errorMessage, ex);
-  }
-
-}
