--- src/main/java/com/student/enterprise/project/repository/GroupRepository.java
@@ -1,8 +1,0 @@
-package com.student.enterprise.project.repository;
-
-import com.student.enterprise.project.entity.GroupEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupRepository extends JpaRepository<GroupEntity, Long> {
-
-}
