--- src/main/java/com/student/enterprise/project/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.student.enterprise.project.repository;
-
-import com.student.enterprise.project.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-}
