+++ src/main/java/student/enterprise/project/SEPApplication.java
@@ -1,0 +1,13 @@
+package student.enterprise.project;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SEPApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SEPApplication.class, args);
+	}
+
+}
