+++ src/main/java/student/enterprise/project/configuration/SepConfig.java
@@ -1,0 +1,8 @@
+package student.enterprise.project.configuration;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SepConfig {
+
+}
