+++ src/main/java/student/enterprise/project/converter/jpa/LocalDateAttributeConverter.java
@@ -1,0 +1,20 @@
+package student.enterprise.project.converter.jpa;
+
+import java.sql.Date;
+import java.time.LocalDate;
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
+
+@Converter
+public class LocalDateAttributeConverter implements AttributeConverter<LocalDate, Date> {
+
+  @Override
+  public Date convertToDatabaseColumn(LocalDate attribute) {
+    return attribute == null ? null : Date.valueOf(attribute);
+  }
+
+  @Override
+  public LocalDate convertToEntityAttribute(Date dbData) {
+    return dbData == null ? null : dbData.toLocalDate();
+  }
+}
