+++ src/main/java/student/enterprise/project/converter/jpa/LocalTimeAttributeConverter.java
@@ -1,0 +1,20 @@
+package student.enterprise.project.converter.jpa;
+
+import java.sql.Time;
+import java.time.LocalTime;
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
+
+@Converter(autoApply = true)
+public class LocalTimeAttributeConverter implements AttributeConverter<LocalTime, Time> {
+
+  @Override
+  public Time convertToDatabaseColumn(LocalTime attribute) {
+    return attribute == null ? null : Time.valueOf(attribute);
+  }
+
+  @Override
+  public LocalTime convertToEntityAttribute(Time dbData) {
+    return dbData == null ? null : dbData.toLocalTime();
+  }
+}
