+++ src/main/java/student/enterprise/project/dto/enums/UserRole.java
@@ -1,0 +1,7 @@
+package student.enterprise.project.dto.enums;
+
+public enum UserRole {
+  USER,
+  MODERATOR,
+  OWNER
+}
