+++ src/main/java/student/enterprise/project/entity/UserGroupEntity.java
@@ -1,0 +1,42 @@
+package student.enterprise.project.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/*
+  Junction Table between users and groups. Also, defines roles etc.
+ */
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "user_group_jt")
+public class UserGroupEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne
+  @JoinColumn(name = "group_id", referencedColumnName = "id")
+  private GroupEntity group;
+
+  @ManyToOne
+  @JoinColumn(name = "user_id", referencedColumnName = "id")
+  private UserEntity user;
+
+  @ManyToOne
+  @JoinColumn(name = "role_id", referencedColumnName = "id")
+  private UserRoleEntity userRole;
+
+}
