+++ src/main/java/student/enterprise/project/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package student.enterprise.project.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import student.enterprise.project.entity.UserEntity;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+}
