--- src/test/java/com/student/enterprise/project/SEPApplicationTests.java
@@ -1,16 +1,0 @@
-package com.student.enterprise.project;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SEPApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
